+++ src/serialization.hpp
@@ -324,7 +324,7 @@ namespace upcxx {
         detail::template memcpy_aligned<alignof(T)>(spot, &x, sizeof(T));
       }
     };
-    
+
     template<bool bounded>
     class serialization_writer;
 
@@ -1313,17 +1313,43 @@ namespace upcxx {
       static typename serialization_traits2<T>::deserialized_type
       deserialized_value(T const &x) {
         UPCXX_ASSERT_INIT();
-        using T1 = typename serialization_traits2<T>::deserialized_type;
+
+        // We access T's serialization through serialization_traits2 instead of
+        // serialization_traits to avoid inheritance circularity since we are
+        // currently in one of serializatoin_traits's base classes. If you are
+        // referencing this code to learn how to manage serialization resources
+        // don't do this, please just use serialization_traits.
+        using the_traits = serialization_traits2<T>;
         
-        auto ub = serialization_traits2<T>::ubound(empty_storage_size, x);
+        using T1 = typename the_traits::deserialized_type;
+
+        auto ub = the_traits::ubound(empty_storage_size, x);
+
+        // Guard against bludgeoning stack with massive static size.
+        // ub::static_size will be massive (-1 or -2) for invalid or dynamic
+        // cases so this unsigned comparison still works.
         constexpr std::size_t static_storage_size = (decltype(ub)::static_size) < 512 ? decltype(ub)::static_size : 512;
-        detail::xaligned_storage<static_storage_size, serialization_align_max> static_storage;
+
+        // Grab some stack. This will be used unless we have a valid bound that
+        // exceeds 512, in which case we heap the bounded buffer. We only need
+        // one but grab an array of 2 so that we can poison the second entry
+        // and catch overflow errors.
+        detail::xaligned_storage<static_storage_size, serialization_align_max> static_storage[2];
         
         void *storage;
         std::size_t storage_size;
-        
-        if(decltype(ub)::is_static || !decltype(ub)::is_valid) {
-          storage = static_storage.storage();
+
+        // We use the stack space if any of the three following conditions hold:
+        // 1. We have size bound within 512.
+        // 2. We have a static size bound within 512. This is implied by the first
+        //    but when true the compiler will be able to prove this branch is
+        //    always taken and prune code.
+        // 3. We have no bound. The stack space is used as the writer's initial hunk.
+        if(/*1*/ ub.size <= 512 ||
+           /*2*/ (decltype(ub)::static_size) <= 512 ||
+           /*3*/ !decltype(ub)::is_valid
+          ) {
+          storage = static_storage[0].storage();
           storage_size = static_storage_size;
         }
         else {
@@ -1332,8 +1358,13 @@ namespace upcxx {
         }
         
         detail::serialization_writer</*bounded=*/decltype(ub)::is_valid> w(storage, storage_size);
-        
-        serialization_traits2<T>::serialize(w, x);
+
+        the_traits::serialize(w, x);
+
+        // Write to space after our static storage to ensure we corrupt any
+        // accidental overflow. This has negligible cost, is only to catch bugs
+        // in testing, and this function probably isn't performance critical anyway.
+        *(char*)static_storage[1].storage() = 0x99;
         
         if(!w.contained_in_initial()) {
           storage_size = w.size();
@@ -1343,9 +1374,9 @@ namespace upcxx {
         
         detail::serialization_reader r(storage);
         detail::raw_storage<T1> x1_raw;
-        (void)serialization_traits2<T>::deserialize(r, &x1_raw);
+        (void)the_traits::deserialize(r, &x1_raw);
         
-        if(storage != static_storage.storage())
+        if(storage != static_storage[0].storage())
           std::free(storage);
         
         return x1_raw.value_and_destruct();